+++ google/config.go
@@ -19,6 +19,7 @@ import (
 	"google.golang.org/api/bigquery/v2"
 	"google.golang.org/api/cloudbilling/v1"
 	"google.golang.org/api/cloudfunctions/v1"
+	"google.golang.org/api/cloudiot/v1"
 	"google.golang.org/api/cloudkms/v1"
 	"google.golang.org/api/cloudresourcemanager/v1"
 	resourceManagerV2Beta1 "google.golang.org/api/cloudresourcemanager/v2beta1"
@@ -47,6 +48,9 @@ type Config struct {
 	Region      string
 	Zone        string
 
+	client    *http.Client
+	userAgent string
+
 	clientBilling                *cloudbilling.Service
 	clientCompute                *compute.Service
 	clientComputeBeta            *computeBeta.Service
@@ -68,6 +72,7 @@ type Config struct {
 	clientServiceMan             *servicemanagement.APIService
 	clientBigQuery               *bigquery.Service
 	clientCloudFunctions         *cloudfunctions.Service
+	clientCloudIoT               *cloudiot.Service
 
 	bigtableClientFactory *BigtableClientFactory
 }
@@ -133,6 +138,9 @@ func (c *Config) loadAndValidate() error {
 	versionString := terraform.VersionString()
 	userAgent := fmt.Sprintf(
 		"(%s %s) Terraform/%s", runtime.GOOS, runtime.GOARCH, versionString)
+
+	c.client = client
+	c.userAgent = userAgent
 
 	var err error
 
@@ -287,6 +295,13 @@ func (c *Config) loadAndValidate() error {
 		return err
 	}
 	c.clientDataproc.UserAgent = userAgent
+
+	log.Printf("[INFO] Instantiating Google Cloud IoT Core Client...")
+	c.clientCloudIoT, err = cloudiot.New(client)
+	if err != nil {
+		return err
+	}
+	c.clientCloudIoT.UserAgent = userAgent
 
 	return nil
 }