+++ CHANGELOG.md
@@ -1,4 +1,15 @@
 ## 1.5.1 (Unreleased)
+
+Features:
+* **New Resource** `google_cloudiot_registry` [GH-970]
+* **New Resource** `google_endpoints_service` [GH-933]
+
+IMPROVEMENTS:
+* compute: Suppress diff if `guest_accelerators` count is 0 in `google_compute_instance` and `google_compute_instance_template` [GH-866]
+* compute: Add update support for machine type, min cpu platform, and service accounts [GH-1005]
+* compute: Add import support for google_compute_shared_vpc_host_project/google_compute_shared_vpc_service_project resources [GH-1004]
+* sql: Support for PostgreSQL high availability [GH-1001]
+
 ## 1.5.0 (January 18, 2018)
 
 FEATURES: